corrección de errores, uso de visual code

+++ src/main/java/com/gl05/bad/dao/BitacoraDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.datatables.repository.DataTablesRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BitacoraDao extends DataTablesRepository<Bitacora, Long>{
+public interface BitacoraDao extends DataTablesRepository<Bitacora, Integer>{
   
 }
